Move Spring Application runner to `runner` package
