+++ scalafmt.conf
@@ -6,5 +6,4 @@ edition = 2019-10
 # Formatter configuration
 # ~~~~~~
 align = more
-docstrings = JavaDoc
 maxColumn = 120
