+++ src/main/scala/com/hhandoko/aktiform/app/runner/AktiformApplication.scala
@@ -1,19 +1,19 @@
-package com.hhandoko.aktiform
+package com.hhandoko.aktiform.app.runner
 
 import org.springframework.boot.SpringApplication
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.context.properties.EnableConfigurationProperties
-import org.springframework.context.annotation.{Configuration, Import}
+import org.springframework.context.annotation.Import
 
 import com.hhandoko.aktiform.app.config.ResourcesConfig
+import com.hhandoko.aktiform.app.controller.{HelloWorldController, HomeController}
 import com.hhandoko.aktiform.app.module.PolyglotContextModule
 
-/**
- * Main application class.
- *
- * `EnableConfigurationProperties` and `Import` is used here to make it explicit configuration and modules that are
- * loaded than relying on automated component scanning.
- */
+/** Main application class (Spring Application runner).
+  *
+  * `EnableConfigurationProperties` and `Import` is used here to make it explicit configuration and modules that are
+  * loaded than relying on automated component scanning.
+  */
 @EnableConfigurationProperties(
   value = Array(
     classOf[ResourcesConfig]
@@ -21,11 +21,14 @@ import com.hhandoko.aktiform.app.module.PolyglotContextModule
 )
 @Import(
   value = Array(
-    classOf[PolyglotContextModule]
+    // Modules
+    classOf[PolyglotContextModule],
+    // Controllers
+    classOf[HelloWorldController],
+    classOf[HomeController]
   )
 )
 @SpringBootApplication
-@Configuration
 class AktiformApplication {}
 
 object AktiformApplication {
